+++ TotalPE/PEStrings.h
@@ -51,7 +51,7 @@ struct PEStrings abstract final {
 	static std::wstring FileFlagsToString(uint32_t flags);
 	static PCWSTR DebugTypeToString(uint32_t type);
 	static PCWSTR CertificateTypeToString(uint32_t type);
-	static PCWSTR x64RelocationTypeToString(BYTE type);
+	static PCWSTR x64RelocationTypeToString(WORD type);
 	static std::wstring CFGFlagsToString(uint32_t flags);
 	static std::wstring GuidToString(GUID const& guid);
 	static PCWSTR SymbolTagToString(SymbolTag tag);
